Added a .sql file to allow users to create the DB schema
We chose not to use Hibernate Auto DDL feature as it does not deal well with a couple of things such as the custom `other` datatype (that powers the integration tests run against H2).

Also, reduced the min idle connection count at the connection pool (Hikari) to avoid exhausting the RDBMS maximum (100 on Postgres) which prevents other connections to be made against the same DB. (e.g. when using a SQL client)